+++ modInfo.json
@@ -2,7 +2,7 @@
   {
     "name" : "scarabol.commands",
     "namepretty" : "Scarabol Commands",
-    "version" : "6.3.14",
+    "version" : "6.3.15",
     "dllpath" : "Commands.dll",
     "enabled" : true,
     "compatibleversions" : [
